Update Run_NVTA_Model.R
This was really interesting to try out!  I made a few changes in lines 71-82 but these can be ignored if need be.  There may be some additional library needed to run the "group_by" and "summarize" command in lines 123-124--that's where I was stopped (R version 4.0.5).  John.Miller@vdot.virginia.gov